Administrate stránka je nyní rozdělená na jednotlivé podstránky
Každá z těchto podstránek má vlastní URL adresu.
Jako domovská administrátorská stránka nyní slouží nový pohled (zatím nenastylovaný) obsahující nějaké základní statistiky.
Stylování většiny elementů je poškozeno v důsledku jejich vyjmutí z původních rodičovských elementů.
Pro každou podstránku není existuje zvláštní kontroler a JavaScript soubor.
CSS soubor je nadále pro všechny podstránky společný a bude rozdělen v budoucích commitech.
Soubor administrate.js je zachován a obsahuje funkci pro předvyplnění formuláře pro odeslání e-mailu. Tato funkce byla upravena tak, aby fungovala i po přesunu formuláře na oddělenou stránku.
