+++ Views/administrate.phtml
@@ -2,33 +2,16 @@
 <div class="container">
     <header>
         <nav>
-            <a id="tab1-link" class="active-tab" onclick="firstTab()">Správa účtů</a>
-            <a id="tab2-link" onclick="secondTab()">Správa tříd</a>
-            <a id="tab3-link" onclick="thirdTab()">Správa hlášení</a>
-            <a id="tab4-link" onclick="fourthTab()">Správa změn jmen</a>
-            <a id="tab5-link" onclick="fifthTab()">Poslat e-mail</a>
-            <a id="tab6-link" onclick="sixthTab()">Ovládání databáze</a>
+            <a href="administrate/users" id="tab1-link">Správa účtů</a>
+            <a href="administrate/classes" id="tab2-link">Správa tříd</a>
+            <a href="administrate/admin-reports" id="tab3-link">Správa hlášení</a>
+            <a href="administrate/name-change-requests" id="tab4-link">Správa změn jmen</a>
+            <a href="administrate/mailsender" id="tab5-link">Poslat e-mail</a>
+            <a href="administrate/database" id="tab6-link">Ovládání databáze</a>
             <a href="menu">Návrat</a>
         </nav>
     </header>
     <main>
-        <div id="tab1">
-            <?php require 'Views/usersTable.phtml'; ?>
-        </div>
-        <div id="tab2">
-            <?php require 'Views/classesTable.phtml'; ?>
-        </div>
-        <div id="tab3">
-            <?php require 'Views/reportsTableAdministrate.phtml'; ?>
-        </div>
-        <div id="tab4">
-            <?php require 'Views/nameChangesTable.phtml'; ?>
-        </div>
-        <div id="tab5">
-            <?php require 'Views/emailForm.phtml'; ?>
-        </div>
-        <div id="tab6">
-            <?php require 'Views/sqlForm.phtml'; ?>
-        </div>
+        <?php require $this->getNextView(); ?> <!-- HLAVNÍ OBSAH STRÁNKY -->
     </main>
 </div>
