+++ Views/stats.phtml
@@ -1,0 +1,58 @@
+<h1>Statistiky</h1>
+<table>
+    <tr>
+        <td>
+            Uživatelé
+        </td>
+        <td>
+            <?= $userCount ?>
+        </td>
+        <td>
+            <a href="administrate/users">Přejít</a>
+        </td>
+    </tr>
+    <tr>
+        <td>
+            Třídy
+        </td>
+        <td>
+            <?= $classesCount ?>
+        </td>
+        <td>
+            <a href="administrate/classes">Přejít</a>
+        </td>
+    </tr>
+    <tr>
+        <td>
+            Administrátorská hlášení
+        </td>
+        <td>
+            <?= $reportCount ?>
+        </td>
+        <td>
+            <a href="administrate/admin-reports">Přejít</a>
+        </td>
+    </tr>
+    <tr>
+        <td>
+            Žádosti o změnu jména účtu
+        </td>
+        <td>
+            <?= $userNameChangeRequestsCount ?>
+        </td>
+        <td>
+            <a href="administrate/name-change-requests">Přejít</a>
+        </td>
+    </tr>
+    <tr>
+        <td>
+            Žádosti o změnu názvu třídy
+        </td>
+        <td>
+            <?= $classNameChangeRequestsCount ?>
+        </td>
+        <td>
+            <a href="administrate/name-change-requests">Přejít</a>
+        </td>
+    </tr>
+</table>
