+++ routes.ini
@@ -14,6 +14,12 @@ recover-password=RecoverPassword
 token-password-change=TokenPasswordChange
 
 administrate=Administrate
+users=Users
+classes=Classes
+admin-reports=AdminReports
+name-change-requests=NameChangeRequests
+mailsender=Mailsender
+database=Database
 administrate-action=AdministrateAction
 
 logout=Logout
@@ -62,6 +68,12 @@ account-update=AccountUpdate
 /token-password-change=TokenPasswordChange
 
 /administrate=Administrate
+/administrate/users=Users
+/administrate/classes=Classes
+/administrate/admin-reports=AdminReports
+/administrate/name-change-requests=NameChangeRequests
+/administrate/mailsender=Mailsender
+/administrate/database=Database
 /administrate-action=AdministrateAction
 /administrate/report-action=ReportAction
 
@@ -145,8 +157,14 @@ account-update=AccountUpdate
 ; - group - musí být zvolena nějaká poznávačka
 ; - naturals - ve zvolené části musí být přítomny nějaké přírodniny (jsou-li zvoleny všechny částí, tak musí být nějaké přírodniny v poznávačce)
 ; - pictures - ve zvolené části musí být přítomny nějaké obrázky (jsou-li zvoleny všechny části, tak musí být nějaké obrázky v poznávačce)
-; Pokud stačí, aby byla splněna pouze jedna z uvedených podmíne (mezi dvěma čárkami), je možné mezi ně vložit symbol "?"
+; Pokud stačí, aby byla splněna pouze jedna z uvedených podmínek (mezi dvěma čárkami), je možné mezi ně vložit symbol "?"
 /administrate=user,systemAdmin
+/administrate/users=user,systemAdmin
+/administrate/classes=user,systemAdmin
+/administrate/admin-reports=user,systemAdmin
+/administrate/name-change-requests=user,systemAdmin
+/administrate/mailsender=user,systemAdmin
+/administrate/database=user,systemAdmin
 /administrate-action=user,systemAdmin
 /report-action=user,systemAdmin?classAdmin
 
@@ -196,7 +214,13 @@ account-update=AccountUpdate
 
 /recover-password/<0>=head,recoverPassword
 
-/administrate=head,administrate
+/administrate=head,administrate,stats
+/users=head,administrate,usersTable
+/classes=head,administrate,classesTable
+/admin-reports=head,administrate,reportsTableAdministrate
+/name-change-requests=head,administrate,nameChangesTable
+/mailsender=head,administrate,emailForm
+/database=head,administrate,sqlForm
 
 /menu=head,menu,menuClassesTable
 /menu/<0>=head,menu,menuGroupsTable
